[I25-187] feat: Lighthouse CI 구성 및 lighthouserc.json 생성 스크립트 추가
