+++ .github/workflows/lighthouse.ci.yml
@@ -14,6 +14,7 @@ jobs:
           node-version: 18
       - run: npm install && npm install -g @lhci/cli@0.15.x
       - run: npm run build
+      - run: node ./generate-lighthouserc.js
       - run: lhci autorun
         env:
           LHCI_GITHUB_APP_TOKEN: ${{ secrets.LHCI_GITHUB_APP_TOKEN }}
