+++ .gitignore
@@ -44,3 +44,4 @@ next-env.d.ts
 
 # lighthouse ci
 .lighthouseci/
+.lighthouserc.json
